Fixed bug that reduced routing performance
When scanning new arcs in Dijkstra's algorithm, it makes sense to check whether the target of the new ark has already been visited and to not scan such arks again.
Up to now however, we always checked if the target of the whole route has already been visited, which hardly makes any sense at all.
Although this typo did not have any impact on the behaviour of the algorithm (the computed routes were still correct) it did have impact on the performance. It was now fixed, and as a consequence the routing algorithms should run faster.