Refactored FullDijkstra and introduced UnreachableNodeProvider
Up to now, FullDijkstra was a standalone implementation of Dijkstra's algorithm which had nothing in common with the other implementations in this project, which all inherit from AbstractDijkstra.

Since this class was only used to compute the unreachable nodes in the graph, it was removed and replaced by UnreachableNodeProvider, which contains a nested implementation of Dijkstra's algorithm, which inherits from PriorityDijkstra.